fixed while.

+++ Scripts/while/while6.py
@@ -1,5 +1,5 @@
-n = float(input())
-fact = 1
+n = int(input())
+fact = 1.0
 
 while n >= 2:
     fact *= n
